[core][windows] restore 'run_check' from 'checks'
PR #1571 introduced a bad regression: 'run_check' method was deleted
beside it is still used on Windows to debug.
